worker/requirements.txt
redis==5.2.1
pymongo==4.10.1
python-dotenv==1.0.1
dnspython==2.7.0
